requirements.txt
# SageForge Production Dependencies for Render Deployment

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Angel One API Integration
smartapi-python==1.3.6
pyotp==2.9.0
requests==2.31.0

# Data Processing & Analysis
pandas==2.1.3
numpy==1.24.4
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# HTTP & CORS
httpx==0.25.2
starlette==0.27.0

# Production Server
gunicorn==21.2.0

# Timezone Support
pytz==2023.3

# Logging & Monitoring
structlog==23.2.0